requirements.txt
python-vlc
tk